sessionStorage added

+++ database/migrations/2019_11_14_034100_create_attendances_table.php
@@ -14,7 +14,7 @@ class CreateAttendancesTable extends Migration
     public function up()
     {
         Schema::create('attendances', function (Blueprint $table) {
-            $table->integer('register_id');
+            $table->bigIncrements('register_id');
             $table->dateTime('starting_date_time');
             $table->dateTime('ending_date_time');
             $table->timestamps();
